+++ tiger-rbel/src/main/java/de/gematik/rbellogger/initializers/RbelKeyFolderInitializer.java
@@ -79,7 +79,10 @@ public class RbelKeyFolderInitializer implements Consumer<RbelConverter> {
     } catch (IOException e) {
       throw new RbelPkiException(
           "Error while initializing keys, failed at file '"
-              + currentFile.get().toAbsolutePath()
+              + Optional.ofNullable(currentFile.get())
+                  .map(Path::toAbsolutePath)
+                  .map(Object::toString)
+                  .orElse("<null>")
               + "'",
           e);
     }